confluence_vector_search/requirements.txt
flask
requests
beautifulsoup4
transformers
torch
faiss-cpu
